+++ api/src/main/java/com/stockManagment/api/produit/ProduitService.java
@@ -1,6 +1,8 @@
 package com.stockManagment.api.produit;
+import com.stockManagment.api.compte.CompteDto;
 import com.stockManagment.api.exceptions.EntityNotFoundException;
 import com.stockManagment.api.exceptions.ErrorCodes;
+import com.stockManagment.api.exceptions.OutOfException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -36,6 +38,39 @@ public class ProduitService {
         if (id == null){
             log.error("Produit non valid !");
         }
-        produitRepo.deleteById(id);
+        if(produitRepo.existsById(id)){
+            ProduitDto produitDto = ProduitDto.fromEntity(produitRepo.findById(id).get());
+            if(produitDto.getQuantiteDisponible() != 0){
+                throw new OutOfException(ErrorCodes.PRODUCT_EXISTS.getDescription(),ErrorCodes.PRODUCT_EXISTS,produitDto.getQuantiteDisponible());
+            }
+            else {
+                produitDto.setIsDeleted(true);
+                produitRepo.save(ProduitDto.toEntity(produitDto));
+            }
+        }
+        else{
+            log.warn("Error ! Product doesn't exist");
+            throw new EntityNotFoundException(ErrorCodes.PRODUIT_NOT_FOUND.getDescription(), ErrorCodes.PRODUIT_NOT_FOUND);
+        }
+    }
+
+    public void deleteById(Integer id) {
+        if(id == null){
+            log.error("Produit id is null");
+        }
+        else{
+            if(produitRepo.existsById(id)){
+                ProduitDto currentProduit = ProduitDto.fromEntity(produitRepo.findById(id).get());
+                if(!currentProduit.getIsDeleted()){
+                    throw new EntityNotFoundException(ErrorCodes.PRODUIT_NOT_FOUND.getDescription() + " logically deleted ! can't delete directly active one",ErrorCodes.PRODUIT_NOT_FOUND);
+                }
+                else{
+                    produitRepo.deleteById(id);
+                }
+            }
+            else {
+                throw new EntityNotFoundException(ErrorCodes.PRODUIT_NOT_FOUND.getDescription(),ErrorCodes.PRODUIT_NOT_FOUND);
+            }
+        }
     }
 }
